[`Materia Stats`] Re-enable
That was a very easy fix... why did it take me so long

+++ Sheets/ExtendedBaseParam.cs
@@ -6,12 +6,12 @@ using Lumina.Excel;
 namespace SimpleTweaksPlugin.Sheets {
     [Sheet( "BaseParam")]
     public class ExtendedBaseParam : Lumina.Excel.GeneratedSheets.BaseParam {
-        public readonly byte[] EquipSlotCategoryPct = new byte[22];
+        public readonly ushort[] EquipSlotCategoryPct = new ushort[22];
 
         public override void PopulateData(RowParser parser, GameData gameData, Language language) {
             base.PopulateData(parser, gameData, language);
             for (var i = 1; i < EquipSlotCategoryPct.Length; i++) {
-                EquipSlotCategoryPct[i] = parser.ReadColumn<byte>(i + 3);
+                EquipSlotCategoryPct[i] = parser.ReadColumn<ushort>(i + 3);
             }
         }
     }
